+++ internal/app/rest_server.go
@@ -2,13 +2,10 @@ package app
 
 import (
 	"context"
-	"encoding/base64"
 	stderrors "errors"
 	"fmt"
-	"maps"
 	"net/http"
 	"runtime"
-	"slices"
 	"sort"
 	"strconv"
 	"strings"
@@ -20,8 +17,6 @@ import (
 	"github.com/go-playground/validator/v10"
 	"github.com/goccy/go-json"
 	"github.com/google/uuid"
-	"github.com/modelcontextprotocol/go-sdk/jsonrpc"
-	mcp "github.com/modelcontextprotocol/go-sdk/mcp"
 	"github.com/rs/zerolog"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protodesc"
@@ -29,7 +24,6 @@ import (
 	"google.golang.org/protobuf/reflect/protoregistry"
 	"google.golang.org/protobuf/types/descriptorpb"
 
-	mcpusecase "github.com/bavix/gripmock/v3/internal/app/usecase/mcp"
 	"github.com/bavix/gripmock/v3/internal/domain/descriptors"
 	"github.com/bavix/gripmock/v3/internal/domain/history"
 	"github.com/bavix/gripmock/v3/internal/domain/rest"
@@ -273,1211 +267,6 @@ func (h *RestServer) DashboardInfo(w http.ResponseWriter, r *http.Request) {
 	})
 }
 
-func (h *RestServer) MCPHandler() http.Handler {
-	h.mcpHandlerOnce.Do(func() {
-		h.mcpHandler = newMCPStreamableHandler(h)
-	})
-
-	return h.mcpHandler
-}
-
-const (
-	debugCallDefaultLimit = 20
-	debugCallHintsCap     = 4
-)
-
-func newMCPStreamableHandler(h *RestServer) http.Handler {
-	server := mcp.NewServer(&mcp.Implementation{Name: "gripmock", Version: build.Version}, nil)
-
-	for _, tool := range mcpusecase.ListRuntimeTools() {
-		name, _ := tool["name"].(string)
-		description, _ := tool["description"].(string)
-		inputSchema, _ := tool["inputSchema"].(map[string]any)
-
-		if name == "" || inputSchema == nil {
-			continue
-		}
-
-		server.AddTool(&mcp.Tool{
-			Name:        name,
-			Description: description,
-			InputSchema: inputSchema,
-		}, newMCPToolHandler(h, name))
-	}
-
-	handler := mcp.NewStreamableHTTPHandler(func(_ *http.Request) *mcp.Server {
-		return server
-	}, &mcp.StreamableHTTPOptions{
-		Stateless:    true,
-		JSONResponse: true,
-	})
-
-	return handler
-}
-
-func newMCPToolHandler(h *RestServer, name string) mcp.ToolHandler {
-	return func(ctx context.Context, req *mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		var args map[string]any
-		if len(req.Params.Arguments) > 0 {
-			if err := json.Unmarshal(req.Params.Arguments, &args); err != nil {
-				return nil, &jsonrpc.Error{Code: jsonrpc.CodeInvalidParams, Message: mcpInvalidArgError("arguments must be an object").Error()}
-			}
-		}
-
-		args = mcpusecase.ApplySession(name, args, mcpSessionFromContext(ctx, req))
-
-		result, err := callMCPToolDispatch(h, name, args)
-		if err != nil {
-			return nil, mcpJSONRPCError(name, err)
-		}
-
-		return &mcp.CallToolResult{
-			Content:           []mcp.Content{&mcp.TextContent{Text: "OK"}},
-			StructuredContent: result,
-		}, nil
-	}
-}
-
-func mcpSessionFromContext(ctx context.Context, req *mcp.CallToolRequest) string {
-	if sessionID := muxmiddleware.FromContext(ctx); sessionID != "" {
-		return sessionID
-	}
-
-	if req == nil || req.Extra == nil {
-		return ""
-	}
-
-	return strings.TrimSpace(req.Extra.Header.Get(muxmiddleware.HeaderName))
-}
-
-func mcpJSONRPCError(toolName string, err error) error {
-	data, marshalErr := json.Marshal(map[string]any{"tool": toolName})
-	if marshalErr != nil {
-		data = nil
-	}
-
-	if stderrors.Is(err, ErrMCPInvalidArgument) {
-		return &jsonrpc.Error{Code: jsonrpc.CodeInvalidParams, Message: err.Error(), Data: data}
-	}
-
-	if stderrors.Is(err, ErrMCPToolNotFound) {
-		return &jsonrpc.Error{Code: jsonrpc.CodeMethodNotFound, Message: err.Error(), Data: data}
-	}
-
-	return &jsonrpc.Error{Code: jsonrpc.CodeInternalError, Message: err.Error(), Data: data}
-}
-
-func callMCPToolDispatch(h *RestServer, name string, args map[string]any) (map[string]any, error) {
-	handlers := mcpToolHandlers(h)
-
-	result, err, found := mcpusecase.DispatchTool(name, args, handlers)
-	if !found {
-		return nil, mcpUnknownTool(name)
-	}
-
-	return result, err
-}
-
-func mcpToolHandlers(h *RestServer) map[string]mcpusecase.ToolHandler {
-	handlers := map[string]mcpusecase.ToolHandler{}
-
-	mergeMCPToolHandlers(handlers, mcpGeneralToolHandlers(h))
-	mergeMCPToolHandlers(handlers, mcpServicesToolHandlers(h))
-	mergeMCPToolHandlers(handlers, mcpStubsToolHandlers(h))
-
-	return handlers
-}
-
-func mcpGeneralToolHandlers(h *RestServer) map[string]mcpusecase.ToolHandler {
-	return map[string]mcpusecase.ToolHandler{
-		mcpusecase.ToolHealthLiveness:  func(toolArgs map[string]any) (map[string]any, error) { return mcpHealthLiveness(h, toolArgs) },
-		mcpusecase.ToolHealthReadiness: func(toolArgs map[string]any) (map[string]any, error) { return mcpHealthReadiness(h, toolArgs) },
-		mcpusecase.ToolHealthStatus:    func(toolArgs map[string]any) (map[string]any, error) { return mcpHealthStatus(h, toolArgs) },
-		mcpusecase.ToolDashboard:       func(toolArgs map[string]any) (map[string]any, error) { return mcpDashboard(h, toolArgs) },
-		mcpusecase.ToolOverview:        func(toolArgs map[string]any) (map[string]any, error) { return mcpDashboardOverview(h, toolArgs) },
-		mcpusecase.ToolInfo:            func(toolArgs map[string]any) (map[string]any, error) { return mcpDashboardInfo(h, toolArgs) },
-		mcpusecase.ToolSessionsList:    func(toolArgs map[string]any) (map[string]any, error) { return mcpSessionsList(h, toolArgs) },
-		mcpusecase.ToolGripmockInfo:    func(toolArgs map[string]any) (map[string]any, error) { return mcpGripmockInfo(h, toolArgs) },
-		mcpusecase.ToolReflectInfo:     func(toolArgs map[string]any) (map[string]any, error) { return mcpReflectInfo(h, toolArgs) },
-		mcpusecase.ToolReflectSources:  func(toolArgs map[string]any) (map[string]any, error) { return mcpReflectSources(h, toolArgs) },
-		mcpusecase.ToolDescriptorsAdd:  func(toolArgs map[string]any) (map[string]any, error) { return mcpDescriptorsAdd(h, toolArgs) },
-		mcpusecase.ToolDescriptorsList: func(toolArgs map[string]any) (map[string]any, error) { return mcpDescriptorsList(h, toolArgs) },
-		mcpusecase.ToolHistoryList:     func(toolArgs map[string]any) (map[string]any, error) { return mcpHistoryList(h, toolArgs) },
-		mcpusecase.ToolHistoryErrors:   func(toolArgs map[string]any) (map[string]any, error) { return mcpHistoryErrors(h, toolArgs) },
-		mcpusecase.ToolVerifyCalls:     func(toolArgs map[string]any) (map[string]any, error) { return mcpVerifyCalls(h, toolArgs) },
-		mcpusecase.ToolDebugCall:       func(toolArgs map[string]any) (map[string]any, error) { return mcpDebugCall(h, toolArgs) },
-		mcpusecase.ToolSchemaStub:      func(toolArgs map[string]any) (map[string]any, error) { return mcpSchemaStub(h, toolArgs) },
-	}
-}
-
-func mcpServicesToolHandlers(h *RestServer) map[string]mcpusecase.ToolHandler {
-	return map[string]mcpusecase.ToolHandler{
-		mcpusecase.ToolServicesList:    func(toolArgs map[string]any) (map[string]any, error) { return mcpServicesList(h, toolArgs) },
-		mcpusecase.ToolServicesGet:     func(toolArgs map[string]any) (map[string]any, error) { return mcpServicesGet(h, toolArgs) },
-		mcpusecase.ToolServicesMethods: func(toolArgs map[string]any) (map[string]any, error) { return mcpServicesMethods(h, toolArgs) },
-		mcpusecase.ToolServicesMethod:  func(toolArgs map[string]any) (map[string]any, error) { return mcpServicesMethod(h, toolArgs) },
-		mcpusecase.ToolServicesDelete:  func(toolArgs map[string]any) (map[string]any, error) { return mcpServicesDelete(h, toolArgs) },
-	}
-}
-
-func mcpStubsToolHandlers(h *RestServer) map[string]mcpusecase.ToolHandler {
-	return map[string]mcpusecase.ToolHandler{
-		mcpusecase.ToolStubsUpsert: func(toolArgs map[string]any) (map[string]any, error) { return mcpStubsUpsert(h, toolArgs) },
-		mcpusecase.ToolStubsList:   func(toolArgs map[string]any) (map[string]any, error) { return mcpStubsList(h, toolArgs) },
-		mcpusecase.ToolStubsGet:    func(toolArgs map[string]any) (map[string]any, error) { return mcpStubsGet(h, toolArgs) },
-		mcpusecase.ToolStubsDelete: func(toolArgs map[string]any) (map[string]any, error) { return mcpStubsDelete(h, toolArgs) },
-		mcpusecase.ToolStubsBatchDelete: func(toolArgs map[string]any) (map[string]any, error) {
-			return mcpStubsBatchDelete(h, toolArgs)
-		},
-		mcpusecase.ToolStubsPurge:   func(toolArgs map[string]any) (map[string]any, error) { return mcpStubsPurge(h, toolArgs) },
-		mcpusecase.ToolStubsSearch:  func(toolArgs map[string]any) (map[string]any, error) { return mcpStubsSearch(h, toolArgs) },
-		mcpusecase.ToolStubsInspect: func(toolArgs map[string]any) (map[string]any, error) { return mcpStubsInspect(h, toolArgs) },
-		mcpusecase.ToolStubsUsed:    func(toolArgs map[string]any) (map[string]any, error) { return mcpStubsUsed(h, toolArgs) },
-		mcpusecase.ToolStubsUnused:  func(toolArgs map[string]any) (map[string]any, error) { return mcpStubsUnused(h, toolArgs) },
-	}
-}
-
-func mergeMCPToolHandlers(dst map[string]mcpusecase.ToolHandler, src map[string]mcpusecase.ToolHandler) {
-	maps.Copy(dst, src)
-}
-
-func mcpSchemaStub(_ *RestServer, _ map[string]any) (map[string]any, error) {
-	return map[string]any{"schemaUrl": stubSchemaURL}, nil
-}
-
-func mcpHealthLiveness(_ *RestServer, _ map[string]any) (map[string]any, error) {
-	return map[string]any{"message": "ok", "time": time.Now()}, nil
-}
-
-func mcpHealthReadiness(h *RestServer, _ map[string]any) (map[string]any, error) {
-	ready := h.ok.Load()
-	if !ready {
-		return map[string]any{"ready": false, "message": "not ready", "time": time.Now()}, nil
-	}
-
-	return map[string]any{"ready": true, "message": "ok", "time": time.Now()}, nil
-}
-
-func mcpHealthStatus(h *RestServer, _ map[string]any) (map[string]any, error) {
-	ready := h.ok.Load()
-
-	readiness := "ok"
-	if !ready {
-		readiness = "not ready"
-	}
-
-	return map[string]any{
-		"liveness":  "ok",
-		"readiness": readiness,
-		"ready":     ready,
-		"time":      time.Now(),
-	}, nil
-}
-
-func mcpDashboard(h *RestServer, args map[string]any) (map[string]any, error) {
-	return map[string]any{"dashboard": h.dashboardPayload(mcpSessionRequest(args))}, nil
-}
-
-func mcpDashboardOverview(h *RestServer, args map[string]any) (map[string]any, error) {
-	payload := h.dashboardPayload(mcpSessionRequest(args))
-
-	return map[string]any{"overview": rest.DashboardOverview{
-		TotalServices:      payload.TotalServices,
-		TotalStubs:         payload.TotalStubs,
-		UsedStubs:          payload.UsedStubs,
-		UnusedStubs:        payload.UnusedStubs,
-		TotalSessions:      payload.TotalSessions,
-		RuntimeDescriptors: payload.RuntimeDescriptors,
-		TotalHistory:       payload.TotalHistory,
-		HistoryErrors:      payload.HistoryErrors,
-	}}, nil
-}
-
-func mcpDashboardInfo(h *RestServer, args map[string]any) (map[string]any, error) {
-	payload := h.dashboardPayload(mcpSessionRequest(args))
-
-	return map[string]any{"info": rest.DashboardInfo{
-		AppName:            payload.AppName,
-		Version:            payload.Version,
-		GoVersion:          payload.GoVersion,
-		Compiler:           payload.Compiler,
-		Goos:               payload.Goos,
-		Goarch:             payload.Goarch,
-		NumCPU:             payload.NumCPU,
-		StartedAt:          payload.StartedAt,
-		UptimeSeconds:      payload.UptimeSeconds,
-		Ready:              payload.Ready,
-		HistoryEnabled:     payload.HistoryEnabled,
-		TotalServices:      payload.TotalServices,
-		TotalStubs:         payload.TotalStubs,
-		TotalSessions:      payload.TotalSessions,
-		RuntimeDescriptors: payload.RuntimeDescriptors,
-	}}, nil
-}
-
-func mcpSessionsList(h *RestServer, _ map[string]any) (map[string]any, error) {
-	return map[string]any{"sessions": h.budgerigar.Sessions()}, nil
-}
-
-func mcpGripmockInfo(h *RestServer, _ map[string]any) (map[string]any, error) {
-	overview := h.dashboardPayload(nil)
-
-	return map[string]any{
-		"appName":            overview.AppName,
-		"version":            overview.Version,
-		"protocolVersion":    mcpusecase.ProtocolVersion,
-		"historyEnabled":     overview.HistoryEnabled,
-		"ready":              overview.Ready,
-		"totalServices":      overview.TotalServices,
-		"totalStubs":         overview.TotalStubs,
-		"totalSessions":      overview.TotalSessions,
-		"runtimeDescriptors": overview.RuntimeDescriptors,
-		"tools":              mcpusecase.ListRuntimeTools(),
-	}, nil
-}
-
-func mcpReflectInfo(h *RestServer, _ map[string]any) (map[string]any, error) {
-	runtimePaths, reflectionPrefixes, reflectionFiles := runtimeDescriptorStats(h)
-
-	globalCount := 0
-
-	protoregistry.GlobalFiles.RangeFiles(func(_ protoreflect.FileDescriptor) bool {
-		globalCount++
-
-		return true
-	})
-
-	return map[string]any{
-		"runtimeDescriptorFiles":    len(runtimePaths),
-		"reflectionDescriptorFiles": reflectionFiles,
-		"dynamicDescriptorFiles":    len(runtimePaths) - reflectionFiles,
-		"reflectionDetected":        reflectionFiles > 0,
-		"reflectionSources":         reflectionPrefixes,
-		"globalDescriptorFiles":     globalCount,
-	}, nil
-}
-
-func mcpReflectSources(h *RestServer, args map[string]any) (map[string]any, error) {
-	runtimePaths, reflectionPrefixes, _ := runtimeDescriptorStats(h)
-	reflectionPaths, dynamicPaths, _ := splitRuntimeDescriptorPaths(runtimePaths)
-
-	kind, _ := args["kind"].(string)
-	if kind == "" {
-		kind = "all"
-	}
-
-	if kind != "all" && kind != "reflection" && kind != "dynamic" {
-		return nil, mcpInvalidArgError("kind must be one of: all, reflection, dynamic")
-	}
-
-	offset, err := mcpIntArg(args, "offset", 0)
-	if err != nil {
-		return nil, err
-	}
-
-	limit, err := mcpIntArg(args, "limit", 0)
-	if err != nil {
-		return nil, err
-	}
-
-	filtered := runtimePaths
-
-	switch kind {
-	case "reflection":
-		filtered = reflectionPaths
-	case "dynamic":
-		filtered = dynamicPaths
-	}
-
-	total := len(filtered)
-	filtered = paginateStringSlice(filtered, offset, limit)
-
-	return map[string]any{
-		"kind":   kind,
-		"paths":  filtered,
-		"total":  total,
-		"offset": offset,
-		"limit":  limit,
-		"groups": map[string]any{
-			"reflection": map[string]any{"count": len(reflectionPaths)},
-			"dynamic":    map[string]any{"count": len(dynamicPaths)},
-		},
-		"reflectionSources": reflectionPrefixes,
-	}, nil
-}
-
-func runtimeDescriptorStats(h *RestServer) ([]string, []string, int) {
-	runtimePaths := h.restDescriptors.Paths()
-	reflectionPaths, _, prefixes := splitRuntimeDescriptorPaths(runtimePaths)
-
-	return runtimePaths, prefixes, len(reflectionPaths)
-}
-
-func splitRuntimeDescriptorPaths(runtimePaths []string) ([]string, []string, []string) {
-	reflectionPrefixes := make(map[string]struct{})
-	reflectionPaths := make([]string, 0, len(runtimePaths))
-	dynamicPaths := make([]string, 0, len(runtimePaths))
-
-	for _, path := range runtimePaths {
-		if !strings.HasPrefix(path, "grpc_reflect_") {
-			dynamicPaths = append(dynamicPaths, path)
-
-			continue
-		}
-
-		reflectionPaths = append(reflectionPaths, path)
-
-		prefix := path
-		if idx := strings.Index(prefix, "/"); idx > 0 {
-			prefix = prefix[:idx]
-		}
-
-		reflectionPrefixes[prefix] = struct{}{}
-	}
-
-	prefixes := make([]string, 0, len(reflectionPrefixes))
-	for prefix := range reflectionPrefixes {
-		prefixes = append(prefixes, prefix)
-	}
-
-	sort.Strings(prefixes)
-
-	return reflectionPaths, dynamicPaths, prefixes
-}
-
-func paginateStringSlice(items []string, offset int, limit int) []string {
-	if offset >= len(items) {
-		return []string{}
-	}
-
-	items = items[offset:]
-
-	if limit > 0 && len(items) > limit {
-		items = items[:limit]
-	}
-
-	return items
-}
-
-func mcpSessionRequest(args map[string]any) *http.Request {
-	req := &http.Request{Header: make(http.Header)}
-	if sessionID, _ := args["session"].(string); sessionID != "" {
-		req.Header.Set(muxmiddleware.HeaderName, sessionID)
-	}
-
-	return req
-}
-
-func mcpDescriptorsAdd(h *RestServer, args map[string]any) (map[string]any, error) {
-	descriptorSetBase64, _ := args["descriptorSetBase64"].(string)
-	if descriptorSetBase64 == "" {
-		return nil, mcpRequiredArgError("descriptorSetBase64")
-	}
-
-	payload, err := base64.StdEncoding.DecodeString(descriptorSetBase64)
-	if err != nil {
-		return nil, mcpDescriptorSetBase64ArgError(err)
-	}
-
-	serviceIDs, err := registerDescriptorBytes(h, payload)
-	if err != nil {
-		return nil, mcpDescriptorRegistrationArgError(err)
-	}
-
-	return map[string]any{"serviceIDs": serviceIDs}, nil
-}
-
-func mcpDescriptorsList(h *RestServer, _ map[string]any) (map[string]any, error) {
-	return map[string]any{"serviceIDs": h.restDescriptors.ServiceIDs()}, nil
-}
-
-func mcpServicesList(h *RestServer, _ map[string]any) (map[string]any, error) {
-	return map[string]any{"services": h.collectAllServices()}, nil
-}
-
-func mcpServicesDelete(h *RestServer, args map[string]any) (map[string]any, error) {
-	serviceID, _ := args["serviceID"].(string)
-	if serviceID == "" {
-		return nil, mcpRequiredArgError("serviceID")
-	}
-
-	removed := unregisterService(h, serviceID)
-
-	return map[string]any{"removed": removed > 0, "serviceID": serviceID}, nil
-}
-
-func mcpServicesGet(h *RestServer, args map[string]any) (map[string]any, error) {
-	serviceID, _ := args["serviceID"].(string)
-	if serviceID == "" {
-		return nil, mcpRequiredArgError("serviceID")
-	}
-
-	service, ok := h.findServiceDetailed(serviceID)
-	if !ok {
-		return nil, mcpInvalidArgError(errServiceNotFound.Error() + ": " + serviceID)
-	}
-
-	return map[string]any{"service": service}, nil
-}
-
-func mcpServicesMethods(h *RestServer, args map[string]any) (map[string]any, error) {
-	serviceID, _ := args["serviceID"].(string)
-	if serviceID == "" {
-		return nil, mcpRequiredArgError("serviceID")
-	}
-
-	serviceDescriptor, ok := h.findServiceDescriptor(serviceID)
-	if !ok {
-		return nil, mcpInvalidArgError(errServiceNotFound.Error() + ": " + serviceID)
-	}
-
-	return map[string]any{"methods": h.serviceFromDescriptor(serviceDescriptor, false).Methods}, nil
-}
-
-func mcpServicesMethod(h *RestServer, args map[string]any) (map[string]any, error) {
-	serviceID, _ := args["serviceID"].(string)
-	if serviceID == "" {
-		return nil, mcpRequiredArgError("serviceID")
-	}
-
-	methodID, _ := args["methodID"].(string)
-	if methodID == "" {
-		return nil, mcpRequiredArgError("methodID")
-	}
-
-	service, ok := h.findServiceDetailed(serviceID)
-	if !ok {
-		return nil, mcpInvalidArgError(errServiceNotFound.Error() + ": " + serviceID)
-	}
-
-	for _, method := range service.Methods {
-		if method.Id == methodID || method.Name == methodID {
-			return map[string]any{"method": method}, nil
-		}
-	}
-
-	return nil, mcpInvalidArgError(errMethodNotFound.Error() + " " + methodID + " in service " + serviceID)
-}
-
-func mcpHistoryList(h *RestServer, args map[string]any) (map[string]any, error) {
-	service, _ := args["service"].(string)
-	method, _ := args["method"].(string)
-	session, _ := args["session"].(string)
-
-	limit, err := mcpIntArg(args, "limit", 0)
-	if err != nil {
-		return nil, err
-	}
-
-	records := filterHistory(h, history.FilterOpts{
-		Service: service,
-		Method:  method,
-		Session: session,
-	}, limit)
-
-	return map[string]any{"records": records}, nil
-}
-
-func mcpHistoryErrors(h *RestServer, args map[string]any) (map[string]any, error) {
-	session, _ := args["session"].(string)
-
-	limit, err := mcpIntArg(args, "limit", 0)
-	if err != nil {
-		return nil, err
-	}
-
-	errorsOnly := extractErrorRecords(filterHistory(h, history.FilterOpts{Session: session}, 0))
-	if limit > 0 && len(errorsOnly) > limit {
-		errorsOnly = errorsOnly[len(errorsOnly)-limit:]
-	}
-
-	return map[string]any{"records": errorsOnly}, nil
-}
-
-func mcpVerifyCalls(h *RestServer, args map[string]any) (map[string]any, error) {
-	service, _ := args["service"].(string)
-	if service == "" {
-		return nil, mcpRequiredArgError("service")
-	}
-
-	method, _ := args["method"].(string)
-	if method == "" {
-		return nil, mcpRequiredArgError("method")
-	}
-
-	expectedCount, err := mcpIntArg(args, "expectedCount", -1)
-	if err != nil {
-		return nil, err
-	}
-
-	if expectedCount < 0 {
-		return nil, mcpRequiredArgError("expectedCount")
-	}
-
-	if h.history == nil {
-		return map[string]any{"verified": false, "message": "history is disabled", "expected": expectedCount, "actual": 0}, nil
-	}
-
-	session, _ := args["session"].(string)
-	calls := h.history.Filter(history.FilterOpts{Service: service, Method: method, Session: session})
-	actual := len(calls)
-
-	if actual != expectedCount {
-		return map[string]any{
-			"verified": false,
-			"message":  fmt.Sprintf("expected %s/%s to be called %d times, got %d", service, method, expectedCount, actual),
-			"expected": expectedCount,
-			"actual":   actual,
-		}, nil
-	}
-
-	return map[string]any{"verified": true, "message": "ok", "expected": expectedCount, "actual": actual}, nil
-}
-
-func mcpDebugCall(h *RestServer, args map[string]any) (map[string]any, error) {
-	service, _ := args["service"].(string)
-	if service == "" {
-		return nil, mcpRequiredArgError("service")
-	}
-
-	method, _ := args["method"].(string)
-	session, _ := args["session"].(string)
-
-	limit, err := mcpIntArg(args, "limit", debugCallDefaultLimit)
-	if err != nil {
-		return nil, err
-	}
-
-	stubsLimit, err := mcpIntArg(args, "stubsLimit", debugCallDefaultLimit)
-	if err != nil {
-		return nil, err
-	}
-
-	return debugCall(h, service, method, session, limit, stubsLimit), nil
-}
-
-func mcpStubsUpsert(h *RestServer, args map[string]any) (map[string]any, error) {
-	rawStubs, ok := args["stubs"]
-	if !ok || rawStubs == nil {
-		return nil, mcpRequiredArgError("stubs")
-	}
-
-	stubs, err := decodeMCPStubsArg(rawStubs)
-	if err != nil {
-		return nil, err
-	}
-
-	sessionID, _ := args["session"].(string)
-
-	for _, stub := range stubs {
-		stub.Session = sessionID
-
-		if err = h.validateStub(stub); err != nil {
-			return nil, mcpInvalidArgErrorWithCause(err.Error(), err)
-		}
-	}
-
-	ids := h.budgerigar.PutMany(stubs...)
-
-	return map[string]any{"ids": uuidListToStringSlice(ids)}, nil
-}
-
-func mcpStubsList(h *RestServer, args map[string]any) (map[string]any, error) {
-	stubs, err := listMCPStubs(h.budgerigar.All(), args)
-	if err != nil {
-		return nil, err
-	}
-
-	return map[string]any{"stubs": stubs}, nil
-}
-
-func mcpStubsUsed(h *RestServer, args map[string]any) (map[string]any, error) {
-	stubs, err := listMCPStubs(h.budgerigar.Used(), args)
-	if err != nil {
-		return nil, err
-	}
-
-	return map[string]any{"stubs": stubs}, nil
-}
-
-func mcpStubsUnused(h *RestServer, args map[string]any) (map[string]any, error) {
-	stubs, err := listMCPStubs(h.budgerigar.Unused(), args)
-	if err != nil {
-		return nil, err
-	}
-
-	return map[string]any{"stubs": stubs}, nil
-}
-
-func mcpStubsGet(h *RestServer, args map[string]any) (map[string]any, error) {
-	id, err := mcpUUIDArg(args, "id")
-	if err != nil {
-		return nil, err
-	}
-
-	found := h.budgerigar.FindByID(id)
-
-	if found == nil {
-		return map[string]any{"found": false, "id": id.String()}, nil
-	}
-
-	return map[string]any{"found": true, "stub": found}, nil
-}
-
-func mcpStubsDelete(h *RestServer, args map[string]any) (map[string]any, error) {
-	id, err := mcpUUIDArg(args, "id")
-	if err != nil {
-		return nil, err
-	}
-
-	deleted := h.budgerigar.DeleteByID(id) > 0
-
-	return map[string]any{"deleted": deleted, "id": id.String()}, nil
-}
-
-func mcpStubsBatchDelete(h *RestServer, args map[string]any) (map[string]any, error) {
-	idStrings, err := mcpStringSliceArg(args, "ids")
-	if err != nil {
-		return nil, err
-	}
-
-	ids := make([]uuid.UUID, 0, len(idStrings))
-	deletedIDs := make([]string, 0, len(idStrings))
-	notFoundIDs := make([]string, 0)
-
-	for _, idString := range idStrings {
-		id, parseErr := uuid.Parse(idString)
-		if parseErr != nil {
-			return nil, mcpUUIDArgError("ids", idString, parseErr)
-		}
-
-		ids = append(ids, id)
-
-		if h.budgerigar.FindByID(id) == nil {
-			notFoundIDs = append(notFoundIDs, idString)
-		} else {
-			deletedIDs = append(deletedIDs, idString)
-		}
-	}
-
-	if len(ids) > 0 {
-		h.budgerigar.DeleteByID(ids...)
-	}
-
-	return map[string]any{
-		"deletedIds":  deletedIDs,
-		"notFoundIds": notFoundIDs,
-	}, nil
-}
-
-func mcpStubsPurge(h *RestServer, args map[string]any) (map[string]any, error) {
-	sessionID, _ := args["session"].(string)
-	if sessionID != "" {
-		deletedCount := h.budgerigar.DeleteSession(sessionID)
-
-		return map[string]any{"deletedCount": deletedCount, "session": sessionID}, nil
-	}
-
-	deletedCount := len(h.budgerigar.All())
-	h.budgerigar.Clear()
-
-	return map[string]any{"deletedCount": deletedCount}, nil
-}
-
-func mcpStubsSearch(h *RestServer, args map[string]any) (map[string]any, error) {
-	service, _ := args["service"].(string)
-	if service == "" {
-		return nil, mcpRequiredArgError("service")
-	}
-
-	method, _ := args["method"].(string)
-	if method == "" {
-		return nil, mcpRequiredArgError("method")
-	}
-
-	input, err := mcpSearchInput(args)
-	if err != nil {
-		return nil, err
-	}
-
-	headers, err := mcpHeadersArg(args)
-	if err != nil {
-		return nil, err
-	}
-
-	sessionID, _ := args["session"].(string)
-
-	result, searchErr := h.budgerigar.FindByQuery(stuber.Query{
-		Service: service,
-		Method:  method,
-		Session: sessionID,
-		Headers: headers,
-		Input:   input,
-	})
-	if searchErr != nil {
-		return mcpSearchNotMatchedResponse(searchErr), nil
-	}
-
-	found := result.Found()
-	if found == nil {
-		response := map[string]any{"matched": false}
-
-		if similar := result.Similar(); similar != nil {
-			response["similarStubId"] = similar.ID.String()
-		}
-
-		return response, nil
-	}
-
-	return map[string]any{
-		"matched": true,
-		"stubId":  found.ID.String(),
-		"output":  found.Output,
-	}, nil
-}
-
-func mcpStubsInspect(h *RestServer, args map[string]any) (map[string]any, error) {
-	query, err := mcpInspectQuery(args)
-	if err != nil {
-		return nil, err
-	}
-
-	report := h.budgerigar.InspectQuery(query)
-
-	return map[string]any{"report": toRestInspectReport(report)}, nil
-}
-
-func mcpInspectQuery(args map[string]any) (stuber.Query, error) {
-	service, _ := args["service"].(string)
-	if service == "" {
-		return stuber.Query{}, mcpRequiredArgError("service")
-	}
-
-	method, _ := args["method"].(string)
-	if method == "" {
-		return stuber.Query{}, mcpRequiredArgError("method")
-	}
-
-	query := stuber.Query{Service: service, Method: method}
-
-	err := mcpInspectQueryOptions(args, &query)
-	if err != nil {
-		return stuber.Query{}, err
-	}
-
-	return query, nil
-}
-
-func mcpInspectQueryOptions(args map[string]any, query *stuber.Query) error {
-	if query == nil {
-		return nil
-	}
-
-	if idValue, ok := args["id"]; ok && idValue != nil {
-		id, err := mcpUUIDArg(args, "id")
-		if err != nil {
-			return err
-		}
-
-		query.ID = &id
-	}
-
-	if sessionID, _ := args["session"].(string); sessionID != "" {
-		query.Session = sessionID
-	}
-
-	headers, err := mcpHeadersArg(args)
-	if err != nil {
-		return err
-	}
-
-	query.Headers = headers
-
-	if rawInput, ok := args["input"]; ok && rawInput != nil {
-		input, err := parseMCPInputArg(rawInput)
-		if err != nil {
-			return err
-		}
-
-		query.Input = input
-	}
-
-	return nil
-}
-
-func decodeMCPStubsArg(raw any) ([]*stuber.Stub, error) {
-	payload, err := json.Marshal(raw)
-	if err != nil {
-		return nil, mcpStubPayloadArgError(err)
-	}
-
-	var items []*stuber.Stub
-	if err = jsondecoder.UnmarshalSlice(payload, &items); err != nil {
-		return nil, mcpStubPayloadArgError(err)
-	}
-
-	if len(items) == 0 {
-		return nil, mcpInvalidArgError("stubs cannot be empty")
-	}
-
-	return items, nil
-}
-
-func listMCPStubs(stubs []*stuber.Stub, args map[string]any) ([]*stuber.Stub, error) {
-	service, _ := args["service"].(string)
-	method, _ := args["method"].(string)
-	sessionID, _ := args["session"].(string)
-
-	limit, err := mcpIntArg(args, "limit", 0)
-	if err != nil {
-		return nil, err
-	}
-
-	offset, err := mcpIntArg(args, "offset", 0)
-	if err != nil {
-		return nil, err
-	}
-
-	filtered := filterMCPStubs(stubs, service, method, sessionID)
-
-	if offset >= len(filtered) {
-		return []*stuber.Stub{}, nil
-	}
-
-	filtered = filtered[offset:]
-
-	if limit > 0 && len(filtered) > limit {
-		filtered = filtered[:limit]
-	}
-
-	return filtered, nil
-}
-
-func mcpUUIDArg(args map[string]any, key string) (uuid.UUID, error) {
-	value, _ := args[key].(string)
-	if value == "" {
-		return uuid.Nil, mcpRequiredArgError(key)
-	}
-
-	id, err := uuid.Parse(value)
-	if err != nil {
-		return uuid.Nil, mcpUUIDArgError(key, value, err)
-	}
-
-	return id, nil
-}
-
-func mcpStringSliceArg(args map[string]any, key string) ([]string, error) {
-	raw, ok := args[key]
-	if !ok || raw == nil {
-		return nil, mcpRequiredArgError(key)
-	}
-
-	switch values := raw.(type) {
-	case []string:
-		return validateMCPStringSlice(values, key)
-	case []any:
-		return convertMCPAnyStringSlice(values, key)
-	default:
-		return nil, mcpStringListArgError(key)
-	}
-}
-
-func mcpSearchInput(args map[string]any) ([]map[string]any, error) {
-	if rawInput, ok := args["input"]; ok && rawInput != nil {
-		return parseMCPInputArg(rawInput)
-	}
-
-	payload, ok := args["payload"].(map[string]any)
-	if !ok || payload == nil {
-		return nil, mcpRequiredArgError("payload")
-	}
-
-	return []map[string]any{payload}, nil
-}
-
-func mcpHeadersArg(args map[string]any) (map[string]any, error) {
-	rawHeaders, ok := args["headers"]
-	if !ok || rawHeaders == nil {
-		return map[string]any{}, nil
-	}
-
-	headers, ok := rawHeaders.(map[string]any)
-	if !ok {
-		return nil, mcpInvalidArgError("headers must be an object")
-	}
-
-	return headers, nil
-}
-
-func mcpSearchNotMatchedResponse(searchErr error) map[string]any {
-	return map[string]any{"matched": false, "error": searchErr.Error()}
-}
-
-func filterMCPStubs(stubs []*stuber.Stub, service, method, sessionID string) []*stuber.Stub {
-	filtered := make([]*stuber.Stub, 0, len(stubs))
-
-	for _, stub := range stubs {
-		if !mcpStubMatchesFilters(stub, service, method, sessionID) {
-			continue
-		}
-
-		filtered = append(filtered, stub)
-	}
-
-	return filtered
-}
-
-func mcpStubMatchesFilters(stub *stuber.Stub, service, method, sessionID string) bool {
-	if service != "" && stub.Service != service {
-		return false
-	}
-
-	if method != "" && stub.Method != method {
-		return false
-	}
-
-	return stubVisibleForSession(stub.Session, sessionID)
-}
-
-func validateMCPStringSlice(values []string, key string) ([]string, error) {
-	if len(values) == 0 {
-		return nil, mcpInvalidArgError(key + " cannot be empty")
-	}
-
-	if slices.Contains(values, "") {
-		return nil, mcpStringListArgError(key)
-	}
-
-	return values, nil
-}
-
-func convertMCPAnyStringSlice(values []any, key string) ([]string, error) {
-	if len(values) == 0 {
-		return nil, mcpInvalidArgError(key + " cannot be empty")
-	}
-
-	out := make([]string, 0, len(values))
-	for _, item := range values {
-		value, ok := item.(string)
-		if !ok || value == "" {
-			return nil, mcpStringListArgError(key)
-		}
-
-		out = append(out, value)
-	}
-
-	return out, nil
-}
-
-func parseMCPInputArg(rawInput any) ([]map[string]any, error) {
-	switch input := rawInput.(type) {
-	case []map[string]any:
-		if len(input) == 0 {
-			return nil, mcpInvalidArgError("input cannot be empty")
-		}
-
-		return input, nil
-	case []any:
-		if len(input) == 0 {
-			return nil, mcpInvalidArgError("input cannot be empty")
-		}
-
-		return convertMCPAnyMapSlice(input)
-	default:
-		return nil, mcpInvalidArgError("input must be an array")
-	}
-}
-
-func convertMCPAnyMapSlice(input []any) ([]map[string]any, error) {
-	out := make([]map[string]any, 0, len(input))
-	for _, item := range input {
-		message, ok := item.(map[string]any)
-		if !ok {
-			return nil, mcpInvalidArgError("input must contain JSON objects")
-		}
-
-		out = append(out, message)
-	}
-
-	return out, nil
-}
-
-func uuidListToStringSlice(ids []uuid.UUID) []string {
-	out := make([]string, 0, len(ids))
-
-	for _, id := range ids {
-		out = append(out, id.String())
-	}
-
-	return out
-}
-
-func debugCall(h *RestServer, service, method, session string, historyLimit, stubsLimit int) map[string]any {
-	serviceFound, methodFound := lookupServiceAndMethod(h, service, method)
-	dynamic := slices.Contains(h.restDescriptors.ServiceIDs(), service)
-	stubCount, stubRecords := collectDebugStubs(h, service, method, session, stubsLimit)
-	historyRecords := filterHistory(h, history.FilterOpts{Service: service, Method: method, Session: session}, historyLimit)
-	errorRecords := extractErrorRecords(historyRecords)
-	hints := buildDebugHints(h, serviceFound, methodFound, method, stubCount)
-
-	return map[string]any{
-		"service":           service,
-		"method":            method,
-		"session":           session,
-		"serviceRegistered": serviceFound,
-		"methodRegistered":  methodFound,
-		"dynamicService":    dynamic,
-		"stubCount":         stubCount,
-		"stubs":             stubRecords,
-		"historyCount":      len(historyRecords),
-		"errorCount":        len(errorRecords),
-		"recentHistory":     historyRecords,
-		"recentErrors":      errorRecords,
-		"hints":             hints,
-	}
-}
-
-func lookupServiceAndMethod(h *RestServer, service, method string) (bool, bool) {
-	for _, svc := range h.collectAllServices() {
-		if svc.Id != service {
-			continue
-		}
-
-		if method == "" {
-			return true, true
-		}
-
-		for _, m := range svc.Methods {
-			if m.Name == method || m.Id == service+"/"+method {
-				return true, true
-			}
-		}
-
-		return true, false
-	}
-
-	return false, false
-}
-
-func collectDebugStubs(h *RestServer, service, method, session string, stubsLimit int) (int, []map[string]any) {
-	stubRecords := make([]map[string]any, 0)
-	stubCount := 0
-
-	for _, stub := range h.budgerigar.All() {
-		if stub.Service != service {
-			continue
-		}
-
-		if !stubVisibleForSession(stub.Session, session) {
-			continue
-		}
-
-		if method != "" && stub.Method != method {
-			continue
-		}
-
-		stubCount++
-
-		if stubsLimit > 0 && len(stubRecords) >= stubsLimit {
-			continue
-		}
-
-		stubRecords = append(stubRecords, map[string]any{
-			"id":       stub.ID.String(),
-			"service":  stub.Service,
-			"method":   stub.Method,
-			"session":  stub.Session,
-			"priority": stub.Priority,
-		})
-	}
-
-	return stubCount, stubRecords
-}
-
-func stubVisibleForSession(stubSession, querySession string) bool {
-	if querySession == "" {
-		return stubSession == ""
-	}
-
-	return stubSession == "" || stubSession == querySession
-}
-
-func extractErrorRecords(records []rest.CallRecord) []rest.CallRecord {
-	errorsOnly := make([]rest.CallRecord, 0)
-
-	for _, item := range records {
-		if item.Error != nil && *item.Error != "" {
-			errorsOnly = append(errorsOnly, item)
-		}
-	}
-
-	return errorsOnly
-}
-
-func buildDebugHints(h *RestServer, serviceFound, methodFound bool, method string, stubCount int) []string {
-	hints := make([]string, 0, debugCallHintsCap)
-
-	if !serviceFound {
-		hints = append(hints, "Service is not registered. Add descriptors first (MCP descriptors.add).")
-	}
-
-	if serviceFound && method != "" && !methodFound {
-		hints = append(hints, "Method is not found in service descriptor.")
-	}
-
-	if serviceFound && methodFound && stubCount == 0 {
-		hints = append(hints, "No stubs found for this service/method. Add one via /api/stubs.")
-	}
-
-	if h.history == nil {
-		hints = append(hints, "History is disabled; enable HISTORY_ENABLED=true to inspect call traces.")
-	}
-
-	return hints
-}
-
-func filterHistory(h *RestServer, opts history.FilterOpts, limit int) []rest.CallRecord {
-	if h.history == nil {
-		return []rest.CallRecord{}
-	}
-
-	calls := h.history.Filter(opts)
-	if limit > 0 && len(calls) > limit {
-		calls = calls[len(calls)-limit:]
-	}
-
-	out := make([]rest.CallRecord, len(calls))
-	for i, c := range calls {
-		out[i] = historyCallRecordToRest(c)
-	}
-
-	return out
-}
-
-func mcpIntArg(args map[string]any, key string, defaultValue int) (int, error) {
-	raw, ok := args[key]
-	if !ok || raw == nil {
-		return defaultValue, nil
-	}
-
-	switch v := raw.(type) {
-	case float64:
-		if v < 0 || v != float64(int(v)) {
-			return 0, mcpNonNegativeIntegerArgError(key)
-		}
-
-		return int(v), nil
-	case int:
-		if v < 0 {
-			return 0, mcpNonNegativeIntegerArgError(key)
-		}
-
-		return v, nil
-	default:
-		return 0, mcpNonNegativeIntegerArgError(key)
-	}
-}
 
 // ListHistory returns recorded gRPC calls.
 func (h *RestServer) ListHistory(w http.ResponseWriter, r *http.Request) {
@@ -1646,10 +435,10 @@ func (h *RestServer) AddDescriptors(w http.ResponseWriter, r *http.Request) {
 
 // DeleteService removes a service added via POST /descriptors.
 // Services from startup (proto path) cannot be removed and return 404.
-func (h *RestServer) DeleteService(w http.ResponseWriter, _ *http.Request, serviceID string) {
+func (h *RestServer) DeleteService(w http.ResponseWriter, r *http.Request, serviceID string) {
 	if unregisterService(h, serviceID) == 0 {
 		w.WriteHeader(http.StatusNotFound)
-		h.writeResponseError(context.Background(), w, serviceNotRemovable(serviceID))
+		h.writeResponseError(r.Context(), w, serviceNotRemovable(serviceID))
 
 		return
 	}
