+++ internal/config/config.go
@@ -1,4 +1,3 @@
-//nolint:tagalign
 package config
 
 import (
@@ -12,26 +11,49 @@ import (
 // ByteSize is kept for backward compatibility; defined in internal/infra/types.
 type ByteSize = infraTypes.ByteSize
 
+// TLSConfig holds TLS settings shared across servers.
+type TLSConfig struct {
+	CertFile   string `env:"CERT_FILE"`
+	KeyFile    string `env:"KEY_FILE"`
+	ClientAuth bool   `env:"CLIENT_AUTH" envDefault:"false"`
+	CAFile     string `env:"CA_FILE"`
+	MinVersion string `env:"MIN_VERSION" envDefault:"1.2"`
+}
+
+// ServerConfig holds address configuration for a server.
+type ServerConfig struct {
+	Host string `env:"HOST" envDefault:"0.0.0.0"`
+	Port string `env:"PORT"`
+	Addr string `env:"ADDR"`
+}
+
+// OTelConfig holds OpenTelemetry configuration.
+type OTelConfig struct {
+	Enabled  bool   `env:"ENABLED" envDefault:"false"`
+	Endpoint string `env:"EXPORTER_OTLP_ENDPOINT" envDefault:"localhost:4317"`
+	Insecure bool   `env:"EXPORTER_OTLP_INSECURE" envDefault:"true"`
+}
+
 // Config holds environment-derived configuration values.
 type Config struct {
 	// Application logging level.
 	LogLevel string `env:"LOG_LEVEL" envDefault:"info"`
 
 	// GRPC server configuration.
-	GRPCNetwork string `env:"GRPC_NETWORK" envDefault:"tcp"`
-	GRPCHost    string `env:"GRPC_HOST" envDefault:"0.0.0.0"`
-	GRPCPort    string `env:"GRPC_PORT" envDefault:"4770"`
-	GRPCAddr    string `env:",expand" envDefault:"$GRPC_HOST:$GRPC_PORT"`
+	GRPCNetwork string     `env:"GRPC_NETWORK" envDefault:"tcp"`
+	GRPC        ServerConfig `envPrefix:"GRPC_"`
+	GRPCTLS     TLSConfig    `envPrefix:"GRPC_TLS_"`
 
 	// HTTP server configuration.
-	HTTPHost string `env:"HTTP_HOST" envDefault:"0.0.0.0"`
-	HTTPPort string `env:"HTTP_PORT" envDefault:"4771"`
-	HTTPAddr string `env:",expand" envDefault:"$HTTP_HOST:$HTTP_PORT"`
+	HTTP    ServerConfig `envPrefix:"HTTP_"`
+	HTTPTLS TLSConfig    `envPrefix:"HTTP_TLS_"`
 
 	// ConnectRPC server configuration.
-	ConnectHost string `env:"CONNECTRPC_HOST" envDefault:"0.0.0.0"`
-	ConnectPort string `env:"CONNECTRPC_PORT" envDefault:"4769"`
-	ConnectAddr string `env:",expand" envDefault:"$CONNECTRPC_HOST:$CONNECTRPC_PORT"`
+	Connect    ServerConfig `envPrefix:"CONNECTRPC_"`
+	ConnectTLS TLSConfig    `envPrefix:"CONNECTRPC_TLS_"`
+
+	// OpenTelemetry configuration.
+	OTel OTelConfig `envPrefix:"OTEL_"`
 
 	// Files configuration.
 	StubWatcherEnabled  bool          `env:"STUB_WATCHER_ENABLED" envDefault:"true"`
@@ -51,30 +73,6 @@ type Config struct {
 	// Plugins configuration.
 	TemplatePluginPaths []string `env:"TEMPLATE_PLUGIN_PATHS"`
 
-	// gRPC TLS configuration.
-	GRPCTLSCertFile   string `env:"GRPC_TLS_CERT_FILE"`
-	GRPCTLSKeyFile    string `env:"GRPC_TLS_KEY_FILE"`
-	GRPCTLSClientAuth bool   `env:"GRPC_TLS_CLIENT_AUTH" envDefault:"false"`
-	GRPCTLSCAFile     string `env:"GRPC_TLS_CA_FILE"`
-	GRPCTLSMinVersion string `env:"GRPC_TLS_MIN_VERSION" envDefault:"1.2"`
-
-	// HTTP TLS configuration.
-	HTTPTLSCertFile   string `env:"HTTP_TLS_CERT_FILE"`
-	HTTPTLSKeyFile    string `env:"HTTP_TLS_KEY_FILE"`
-	HTTPTLSClientAuth bool   `env:"HTTP_TLS_CLIENT_AUTH" envDefault:"false"`
-	HTTPTLSCAFile     string `env:"HTTP_TLS_CA_FILE"`
-
-	// ConnectRPC TLS configuration.
-	ConnectTLSCertFile   string `env:"CONNECTRPC_TLS_CERT_FILE"`
-	ConnectTLSKeyFile    string `env:"CONNECTRPC_TLS_KEY_FILE"`
-	ConnectTLSClientAuth bool   `env:"CONNECTRPC_TLS_CLIENT_AUTH" envDefault:"false"`
-	ConnectTLSCAFile     string `env:"CONNECTRPC_TLS_CA_FILE"`
-
-	// OpenTelemetry configuration.
-	OtelEnabled  bool   `env:"OTEL_ENABLED" envDefault:"false"`
-	OtelEndpoint string `env:"OTEL_EXPORTER_OTLP_ENDPOINT" envDefault:"localhost:4317"`
-	OtelInsecure bool   `env:"OTEL_EXPORTER_OTLP_INSECURE" envDefault:"true"`
-
 	// Buf Schema Registry configuration.
 	BSR BSRConfig `envPrefix:"BSR_"`
 }
@@ -83,7 +81,29 @@ type Config struct {
 func Load() Config {
 	var cfg Config
 
-	_ = env.Parse(&cfg)
+	if err := env.Parse(&cfg); err != nil {
+		panic("config: failed to parse environment: " + err.Error())
+	}
+
+	if cfg.GRPC.Port == "" {
+		cfg.GRPC.Port = "4770"
+	}
+	if cfg.HTTP.Port == "" {
+		cfg.HTTP.Port = "4771"
+	}
+	if cfg.Connect.Port == "" {
+		cfg.Connect.Port = "4769"
+	}
+
+	if cfg.GRPC.Addr == "" {
+		cfg.GRPC.Addr = cfg.GRPC.Host + ":" + cfg.GRPC.Port
+	}
+	if cfg.HTTP.Addr == "" {
+		cfg.HTTP.Addr = cfg.HTTP.Host + ":" + cfg.HTTP.Port
+	}
+	if cfg.Connect.Addr == "" {
+		cfg.Connect.Addr = cfg.Connect.Host + ":" + cfg.Connect.Port
+	}
 
 	return cfg
 }
