+++ internal/config/config_test.go
@@ -28,12 +28,8 @@ func TestConfigOldEnvVarNames(t *testing.T) {
 	expected := config.Config{
 		LogLevel:               "debug",
 		GRPCNetwork:            "tcp",
-		GRPCHost:               "127.0.0.1",
-		GRPCPort:               "8080",
-		GRPCAddr:               "127.0.0.1:8080",
-		HTTPHost:               "localhost",
-		HTTPPort:               "8081",
-		HTTPAddr:               "localhost:8081",
+		GRPC:                   config.ServerConfig{Host: "127.0.0.1", Port: "8080", Addr: "127.0.0.1:8080"},
+		HTTP:                   config.ServerConfig{Host: "localhost", Port: "8081", Addr: "localhost:8081"},
 		StubWatcherEnabled:     false,
 		StubWatcherInterval:    5 * 1000000000, // 5s in nanoseconds
 		StubWatcherType:        "polling",
@@ -50,12 +46,12 @@ func TestConfigOldEnvVarNames(t *testing.T) {
 	// Assert values
 	require.Equal(t, expected.LogLevel, cfg.LogLevel)
 	require.Equal(t, expected.GRPCNetwork, cfg.GRPCNetwork)
-	require.Equal(t, expected.GRPCHost, cfg.GRPCHost)
-	require.Equal(t, expected.GRPCPort, cfg.GRPCPort)
-	require.Equal(t, expected.GRPCAddr, cfg.GRPCAddr)
-	require.Equal(t, expected.HTTPHost, cfg.HTTPHost)
-	require.Equal(t, expected.HTTPPort, cfg.HTTPPort)
-	require.Equal(t, expected.HTTPAddr, cfg.HTTPAddr)
+	require.Equal(t, expected.GRPC.Host, cfg.GRPC.Host)
+	require.Equal(t, expected.GRPC.Port, cfg.GRPC.Port)
+	require.Equal(t, expected.GRPC.Addr, cfg.GRPC.Addr)
+	require.Equal(t, expected.HTTP.Host, cfg.HTTP.Host)
+	require.Equal(t, expected.HTTP.Port, cfg.HTTP.Port)
+	require.Equal(t, expected.HTTP.Addr, cfg.HTTP.Addr)
 	require.Equal(t, expected.StubWatcherEnabled, cfg.StubWatcherEnabled)
 	require.Equal(t, expected.StubWatcherInterval, cfg.StubWatcherInterval)
 	require.Equal(t, expected.StubWatcherType, cfg.StubWatcherType)
@@ -71,12 +67,8 @@ func TestConfigDefaultValues(t *testing.T) {
 	expected := config.Config{
 		LogLevel:               "info",
 		GRPCNetwork:            "tcp",
-		GRPCHost:               "0.0.0.0",
-		GRPCPort:               "4770",
-		GRPCAddr:               "0.0.0.0:4770",
-		HTTPHost:               "0.0.0.0",
-		HTTPPort:               "4771",
-		HTTPAddr:               "0.0.0.0:4771",
+		GRPC:                   config.ServerConfig{Host: "0.0.0.0", Port: "4770", Addr: "0.0.0.0:4770"},
+		HTTP:                   config.ServerConfig{Host: "0.0.0.0", Port: "4771", Addr: "0.0.0.0:4771"},
 		StubWatcherEnabled:     true,
 		StubWatcherInterval:    1 * 1000000000, // 1s in nanoseconds
 		StubWatcherType:        "fsnotify",
@@ -93,12 +85,12 @@ func TestConfigDefaultValues(t *testing.T) {
 	// Assert values
 	require.Equal(t, expected.LogLevel, cfg.LogLevel)
 	require.Equal(t, expected.GRPCNetwork, cfg.GRPCNetwork)
-	require.Equal(t, expected.GRPCHost, cfg.GRPCHost)
-	require.Equal(t, expected.GRPCPort, cfg.GRPCPort)
-	require.Equal(t, expected.GRPCAddr, cfg.GRPCAddr)
-	require.Equal(t, expected.HTTPHost, cfg.HTTPHost)
-	require.Equal(t, expected.HTTPPort, cfg.HTTPPort)
-	require.Equal(t, expected.HTTPAddr, cfg.HTTPAddr)
+	require.Equal(t, expected.GRPC.Host, cfg.GRPC.Host)
+	require.Equal(t, expected.GRPC.Port, cfg.GRPC.Port)
+	require.Equal(t, expected.GRPC.Addr, cfg.GRPC.Addr)
+	require.Equal(t, expected.HTTP.Host, cfg.HTTP.Host)
+	require.Equal(t, expected.HTTP.Port, cfg.HTTP.Port)
+	require.Equal(t, expected.HTTP.Addr, cfg.HTTP.Addr)
 	require.Equal(t, expected.StubWatcherEnabled, cfg.StubWatcherEnabled)
 	require.Equal(t, expected.StubWatcherInterval, cfg.StubWatcherInterval)
 	require.Equal(t, expected.StubWatcherType, cfg.StubWatcherType)
