+++ internal/infra/plugins/wrap_internal_test.go
@@ -9,6 +9,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	pkgplugins "github.com/bavix/gripmock/v3/pkg/plugins"
+
+	"github.com/bavix/gripmock/v3/internal/infra/funcwrap"
 )
 
 const (
@@ -226,7 +228,7 @@ func TestIsNilAssignable(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			require.Equal(t, tt.want, isNilAssignable(tt.typ))
+			require.Equal(t, tt.want, funcwrap.IsNilAssignable(tt.typ))
 		})
 	}
 }
@@ -241,7 +243,7 @@ func TestCoerceArg(t *testing.T) {
 	fnType := reflect.TypeFor[func(int)]()
 
 	// Act
-	val, err := coerceArg(args, &idx, paramType, fnType, 0)
+	val, err := funcwrap.CoerceArg(args, &idx, paramType, fnType, 0)
 
 	// Assert
 	require.NoError(t, err)
@@ -259,7 +261,7 @@ func TestCoerceArgNilValue(t *testing.T) {
 	fnType := reflect.TypeFor[func(*int)]()
 
 	// Act
-	val, err := coerceArg(args, &idx, paramType, fnType, 0)
+	val, err := funcwrap.CoerceArg(args, &idx, paramType, fnType, 0)
 
 	// Assert
 	require.NoError(t, err)
@@ -276,7 +278,7 @@ func TestCoerceArgNotEnoughArgs(t *testing.T) {
 	fnType := reflect.TypeFor[func(int)]()
 
 	// Act
-	_, err := coerceArg(args, &idx, paramType, fnType, 0)
+	_, err := funcwrap.CoerceArg(args, &idx, paramType, fnType, 0)
 
 	// Assert
 	require.Error(t, err)
@@ -293,7 +295,7 @@ func TestCoerceArgTypeMismatch(t *testing.T) {
 	fnType := reflect.TypeFor[func(int)]()
 
 	// Act
-	_, err := coerceArg(args, &idx, paramType, fnType, 0)
+	_, err := funcwrap.CoerceArg(args, &idx, paramType, fnType, 0)
 
 	// Assert
 	require.Error(t, err)
